account: [22.14] Add condition to check if a line is reconciled before attempting to reconcile it when posting a move [CUSTOM]
Fix #PJAZZ-703

When cancelling a move through the move cancellation wizard, we reconcile the lines and then post the generated cancellation move. However, when posting the move, we currently try to reconcile lines on which there is no credit/debit, which results in an error because these lines have already been reconciled in the previous step.